+++ tests/testthat/test-textmodel.R
@@ -112,24 +112,24 @@ test_that("predict.textmodel_lss is working", {
     expect_equal(length(pred3$density), ndoc(toks_test))
     expect_null(names(pred3$density))
 
-    pred4 <- predict(lss_test, rescaling = FALSE)
+    pred4 <- predict(lss_test, rescale = FALSE)
     expect_identical(names(pred4), docnames(toks_test))
     expect_equal(as.numeric(scale(pred4)), unname(pred1))
 
     pred5 <- predict(lss_test, se_fit = TRUE, density = TRUE)
     expect_equal(names(pred5), c("fit", "se.fit", "n", "density"))
 
-    pred6 <- predict(lss_test, rescaling = FALSE, min_n = 2)
+    pred6 <- predict(lss_test, rescale = FALSE, min_n = 2)
     expect_true(all(is.na(pred4) == is.na(pred6)))
     expect_true(all(abs(pred6[pred5$n == 1]) < abs(pred4[pred5$n == 1]), na.rm = TRUE))
     expect_true(all(abs(pred6[pred5$n >= 2]) == abs(pred4[pred5$n >= 2]), na.rm = TRUE))
 
     expect_error(
-        predict(lss_test, rescaling = FALSE, min_n = -1),
+        predict(lss_test, rescale = FALSE, min_n = -1),
         "The value of min_n must be between 0 and Inf"
     )
     expect_error(
-        predict(lss_test, rescaling = FALSE, min_n = c(0, 1)),
+        predict(lss_test, rescale = FALSE, min_n = c(0, 1)),
         "The length of min_n must be 1"
     )
 })
@@ -161,7 +161,7 @@ test_that("calculation of fit and se_fit are correct", {
     lss <- as.textmodel_lss(c("a" = 0.1, "b" = 0.1, "c" = 0.3))
     toks <- tokens(c("a a a", "a b", "a a b c c d e"))
     dfmt <- dfm(toks)
-    pred <- predict(lss, newdata = dfmt, se_fit = TRUE, rescaling = FALSE)
+    pred <- predict(lss, newdata = dfmt, se_fit = TRUE, rescale = FALSE)
 
     expect_equal(pred$fit[1], c(text1 = 0.10))
     expect_equal(pred$fit[2], c(text2 = 0.10))
@@ -434,26 +434,26 @@ test_that("se_fit is working", {
     lss <- as.textmodel_lss(beta)
     dfmt1 <- dfm(tokens(c("a a a b b", "")))
     dfmt2 <- dfm(tokens(c("a a a b b z z z z z", "")))
-    pred1 <- predict(lss, newdata = dfmt1, rescaling = FALSE, min_n = 10, se_fit = TRUE)
-    pred2 <- predict(lss, newdata = dfmt2, rescaling = FALSE, se_fit = TRUE)
+    pred1 <- predict(lss, newdata = dfmt1, rescale = FALSE, min_n = 10, se_fit = TRUE)
+    pred2 <- predict(lss, newdata = dfmt2, rescale = FALSE, se_fit = TRUE)
     expect_identical(pred1, pred2)
 })
 
 test_that("devide is working", {
 
-    p1 <- predict(lss_test, devide = 0.5, rescaling = TRUE)
+    p1 <- predict(lss_test, devide = 0.5, rescale = TRUE)
     expect_true(min(p1, na.rm = TRUE) < -1)
     expect_true(max(p1, na.rm = TRUE) > 1)
 
-    p2 <- predict(lss_test, devide = 0.5, rescaling = FALSE)
+    p2 <- predict(lss_test, devide = 0.5, rescale = FALSE)
     expect_true(min(p2, na.rm = TRUE) >= -0.5)
     expect_true(max(p2, na.rm = TRUE) <= 0.5)
 
-    p3 <- predict(lss_test, devide = 0.5, rescaling = FALSE, min_n = 10)
+    p3 <- predict(lss_test, devide = 0.5, rescale = FALSE, min_n = 10)
     expect_true(min(p3, na.rm = TRUE) >= -0.5)
     expect_true(max(p3, na.rm = TRUE) <= 0.5)
 
-    p4 <- predict(lss_test, devide = 0.90, rescaling = FALSE, min_n = 10)
+    p4 <- predict(lss_test, devide = 0.90, rescale = FALSE, min_n = 10)
     expect_true(min(p4, na.rm = TRUE) >= -0.5)
     expect_true(max(p4, na.rm = TRUE) <= 0.5)
 
@@ -465,4 +465,15 @@ test_that("devide is working", {
         predict(lss_test, devide = c(0.1, 0.5)),
         "The length of devide must be 1"
     )
+})
+
+test_that("rescaling still works", {
+
+    expect_warning({
+        p1 <- predict(lss_test, rescaling = TRUE)
+    })
+    expect_silent({
+        p2 <- predict(lss_test, rescale = TRUE)
+    })
+    expect_identical(p1, p2)
 })